mcp: Move screen/window tools to computercontroller module
Move list_windows and screen_capture tools from the developer MCP
to computercontroller. This is about making the developer MCP more focused.

xref https://github.com/block/goose/issues/2830
