+++ crates/goose-mcp/src/developer/mod.rs
@@ -44,7 +44,6 @@ use self::shell::{expand_path, get_shell_config, is_absolute_path, normalize_lin
 use indoc::indoc;
 use std::process::Stdio;
 use std::sync::{Arc, Mutex};
-use xcap::{Monitor, Window};
 
 use ignore::gitignore::{Gitignore, GitignoreBuilder};
 
@@ -277,61 +276,6 @@ impl DeveloperRouter {
             }),
         );
 
-        let list_windows_tool = Tool::new(
-            "list_windows",
-            indoc! {r#"
-                List all available window titles that can be used with screen_capture.
-                Returns a list of window titles that can be used with the window_title parameter
-                of the screen_capture tool.
-            "#},
-            object!({
-                "type": "object",
-                "required": [],
-                "properties": {}
-            }),
-        )
-        .annotate(ToolAnnotations {
-            title: Some("List available windows".to_string()),
-            read_only_hint: Some(true),
-            destructive_hint: Some(false),
-            idempotent_hint: Some(false),
-            open_world_hint: Some(false),
-        });
-
-        let screen_capture_tool = Tool::new(
-            "screen_capture",
-            indoc! {r#"
-                Capture a screenshot of a specified display or window.
-                You can capture either:
-                1. A full display (monitor) using the display parameter
-                2. A specific window by its title using the window_title parameter
-
-                Only one of display or window_title should be specified.
-            "#},
-            object!({
-                "type": "object",
-                "required": [],
-                "properties": {
-                    "display": {
-                        "type": "integer",
-                        "default": 0,
-                        "description": "The display number to capture (0 is main display)"
-                    },
-                    "window_title": {
-                        "type": "string",
-                        "default": null,
-                        "description": "Optional: the exact title of the window to capture. use the list_windows tool to find the available windows."
-                    }
-                }
-            })
-        ).annotate(ToolAnnotations {
-            title: Some("Capture a full screen".to_string()),
-            read_only_hint: Some(true),
-            destructive_hint: Some(false),
-            idempotent_hint: Some(false),
-            open_world_hint: Some(false),
-        });
-
         let image_processor_tool = Tool::new(
             "image_processor",
             indoc! {r#"
@@ -375,9 +319,6 @@ impl DeveloperRouter {
 
                 Use the shell tool as needed to locate files or interact with the project.
 
-                Your windows/screen tools can be used for visual debugging. You should not use these tools unless
-                prompted to, but you can mention they are available if they are relevant.
-
                 operating system: {os}
                 current directory: {cwd}
 
@@ -391,9 +332,6 @@ impl DeveloperRouter {
 
             You can use the shell tool to run any command that would work on the relevant operating system.
             Use the shell tool as needed to locate files or interact with the project.
-
-            Your windows/screen tools can be used for visual debugging. You should not use these tools unless
-            prompted to, but you can mention they are available if they are relevant.
 
             operating system: {os}
             current directory: {cwd}
@@ -515,13 +453,7 @@ impl DeveloperRouter {
         let ignore_patterns = builder.build().expect("Failed to build ignore patterns");
 
         Self {
-            tools: vec![
-                bash_tool,
-                text_editor_tool,
-                list_windows_tool,
-                screen_capture_tool,
-                image_processor_tool,
-            ],
+            tools: vec![bash_tool, text_editor_tool, image_processor_tool],
             prompts: Arc::new(load_prompt_files()),
             instructions,
             file_history: Arc::new(Mutex::new(HashMap::new())),
@@ -1257,22 +1189,6 @@ impl DeveloperRouter {
         Ok(())
     }
 
-    async fn list_windows(&self, _params: Value) -> Result<Vec<Content>, ToolError> {
-        let windows = Window::all()
-            .map_err(|_| ToolError::ExecutionError("Failed to list windows".into()))?;
-
-        let window_titles: Vec<String> =
-            windows.into_iter().map(|w| w.title().to_string()).collect();
-
-        Ok(vec![
-            Content::text(format!("Available windows:\n{}", window_titles.join("\n")))
-                .with_audience(vec![Role::Assistant]),
-            Content::text(format!("Available windows:\n{}", window_titles.join("\n")))
-                .with_audience(vec![Role::User])
-                .with_priority(0.0),
-        ])
-    }
-
     // Helper function to handle Mac screenshot filenames that contain U+202F (narrow no-break space)
     fn normalize_mac_screenshot_path(&self, path: &Path) -> PathBuf {
         // Only process if the path has a filename
@@ -1391,78 +1307,6 @@ impl DeveloperRouter {
             Content::image(data, "image/png").with_priority(0.0),
         ])
     }
-
-    async fn screen_capture(&self, params: Value) -> Result<Vec<Content>, ToolError> {
-        let mut image = if let Some(window_title) =
-            params.get("window_title").and_then(|v| v.as_str())
-        {
-            // Try to find and capture the specified window
-            let windows = Window::all()
-                .map_err(|_| ToolError::ExecutionError("Failed to list windows".into()))?;
-
-            let window = windows
-                .into_iter()
-                .find(|w| w.title() == window_title)
-                .ok_or_else(|| {
-                    ToolError::ExecutionError(format!(
-                        "No window found with title '{}'",
-                        window_title
-                    ))
-                })?;
-
-            window.capture_image().map_err(|e| {
-                ToolError::ExecutionError(format!(
-                    "Failed to capture window '{}': {}",
-                    window_title, e
-                ))
-            })?
-        } else {
-            // Default to display capture if no window title is specified
-            let display = params.get("display").and_then(|v| v.as_u64()).unwrap_or(0) as usize;
-
-            let monitors = Monitor::all()
-                .map_err(|_| ToolError::ExecutionError("Failed to access monitors".into()))?;
-            let monitor = monitors.get(display).ok_or_else(|| {
-                ToolError::ExecutionError(format!(
-                    "{} was not an available monitor, {} found.",
-                    display,
-                    monitors.len()
-                ))
-            })?;
-
-            monitor.capture_image().map_err(|e| {
-                ToolError::ExecutionError(format!("Failed to capture display {}: {}", display, e))
-            })?
-        };
-
-        // Resize the image to a reasonable width while maintaining aspect ratio
-        let max_width = 768;
-        if image.width() > max_width {
-            let scale = max_width as f32 / image.width() as f32;
-            let new_height = (image.height() as f32 * scale) as u32;
-            image = xcap::image::imageops::resize(
-                &image,
-                max_width,
-                new_height,
-                xcap::image::imageops::FilterType::Lanczos3,
-            )
-        };
-
-        let mut bytes: Vec<u8> = Vec::new();
-        image
-            .write_to(&mut Cursor::new(&mut bytes), xcap::image::ImageFormat::Png)
-            .map_err(|e| {
-                ToolError::ExecutionError(format!("Failed to write image buffer {}", e))
-            })?;
-
-        // Convert to base64
-        let data = base64::prelude::BASE64_STANDARD.encode(bytes);
-
-        Ok(vec![
-            Content::text("Screenshot captured").with_audience(vec![Role::Assistant]),
-            Content::image(data, "image/png").with_priority(0.0),
-        ])
-    }
 }
 
 fn recommend_read_range(path: &Path, total_lines: usize) -> Result<Vec<Content>, ToolError> {
@@ -1506,8 +1350,6 @@ impl Router for DeveloperRouter {
             match tool_name.as_str() {
                 "shell" => this.bash(arguments, notifier).await,
                 "text_editor" => this.text_editor(arguments).await,
-                "list_windows" => this.list_windows(arguments).await,
-                "screen_capture" => this.screen_capture(arguments).await,
                 "image_processor" => this.image_processor(arguments).await,
                 _ => Err(ToolError::NotFound(format!("Tool {} not found", tool_name))),
             }
